Moved configs to nems_web/configs and nems_db/configs, working on re-organizing xfspec plugins in nems_lbhb.
