+++ .gitignore
@@ -208,7 +208,9 @@ Session.vim
 tags
 
 # configurations
-nems_db_configs/*
-nems_db_configs/db_defaults.py
-nems_db_configs/web_defaults.py
-nems_db_configs/__init__.py
+nems_db/configs/*
+!nems_db/configs/db_defaults.py
+!nems_db/configs/__init__.py
+nems_web/configs/*
+!nems_web/configs/web_defaults.py
+!nems_db/configs/__init__.py
