+++ nems_db/__init__.py
@@ -10,7 +10,7 @@ log = logging.getLogger(__name__)
 
 def load_config():
     import os
-    from nems_db_configs import db_defaults
+    from nems_db.configs import db_defaults
 
     # leave off 'db_defaults.py' at end
     configs_path = os.path.dirname(os.path.abspath(db_defaults.__file__))
@@ -19,7 +19,7 @@ def load_config():
     # otherwise, use an empty dictionary and create a blank file
     # to point user to the right place.
     try:
-        from nems_db_configs import db_settings
+        from nems_db.configs import db_settings
     except ImportError:
         db_path = os.path.join(configs_path, 'db_settings.py')
         # this should be equivalent to `touch path/to/configs/db_settings.py`
@@ -27,7 +27,7 @@ def load_config():
             os.utime(db_path, None)
         log.info("No db_settings.py found in configs directory,"
                  " generating blank file ... ")
-        from nems_db_configs import db_settings
+        from nems_db.configs import db_settings
 
     cached_config = {}
     _init_settings(os.environ, db_defaults, db_settings, cached_config)
