+++ nems_web/__init__.py
@@ -10,7 +10,7 @@ log = logging.getLogger(__name__)
 
 def load_config():
     import os
-    from nems_db_configs import web_defaults
+    from nems_web.configs import web_defaults
 
     # leave off 'web_defaults.py' at end
     configs_path = os.path.dirname(os.path.abspath(web_defaults.__file__))
@@ -19,7 +19,7 @@ def load_config():
     # otherwise, use an empty dictionary and create a blank file
     # to point user to the right place.
     try:
-        from nems_db_configs import web_settings
+        from nems_web.configs import web_settings
     except ImportError:
         web_path = os.path.join(configs_path, 'web_settings.py')
         # this should be equivalent to `touch path/to/configs/db_settings.py`
@@ -27,7 +27,7 @@ def load_config():
             os.utime(web_path, None)
         log.info("No web_settings.py found in configs directory,"
                  " generating blank file ... ")
-        from nems_db_configs import web_settings
+        from nems_web.configs import web_settings
 
     cached_config = {}
     _init_settings(os.environ, web_defaults, web_settings, cached_config)
